[MOVE] user -> response, request 패키지 생성 후 이동
