+++ src/main/java/com/modugarden/domain/user/dto/response/UserInfoResponseDto.java
@@ -1,4 +1,4 @@
-package com.modugarden.domain.user.dto;
+package com.modugarden.domain.user.dto.response;
 
 import com.modugarden.domain.user.entity.enums.UserAuthority;
 import lombok.AllArgsConstructor;
