+++ src/main/java/com/modugarden/domain/user/dto/response/UserNicknameResponseDto.java
@@ -1,4 +1,4 @@
-package com.modugarden.domain.user.dto;
+package com.modugarden.domain.user.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
